update delete plan service

+++ src/services/plan/DeletePlanService.ts
@@ -1,5 +1,6 @@
 import CustomError from "../../errors/CustomError";
 import PlanModel from "../../models/PlanModel";
+import SubscriptionModel from "../../models/SubscriptionModel";
 import isNotObjectId from "../../utils/isNotObjectId";
 
 
@@ -10,6 +11,14 @@ const DeletePlanService = async (planId: string) => {
     const plan = await PlanModel.findById(planId)
     if(!plan){
         throw new CustomError(404, 'Plan not found with the provided ID');
+    }
+
+    //check if planId is associated with subscription
+    const associatedWitSubscription= await SubscriptionModel.findOne({
+        planId
+    });
+    if (associatedWitSubscription) {
+        throw new CustomError(409, 'Unable to delete, This Plan is associated with one or more subscriptions');
     }
 
     const result = await PlanModel.deleteOne({ _id: planId})
